Fix time columns

+++ internal/storage/mysql/create.sql
@@ -9,8 +9,8 @@ CREATE TABLE `urls` (
   `id` BIGINT UNSIGNED NOT NULL,
   `url` VARCHAR(255) CHARACTER SET 'utf8mb4' COLLATE 'utf8mb4_0900_ai_ci' NOT NULL,
   `parsed_url` VARCHAR(255) CHARACTER SET 'utf8mb4' COLLATE 'utf8mb4_0900_ai_ci' NOT NULL,
-  `fetch_time` INT NOT NULL,
-  `expires` INT NOT NULL,
+  `fetch_time` BIGINT NOT NULL,
+  `expires` BIGINT NOT NULL,
   `metadata` JSON NOT NULL,
   `content_text` TEXT CHARACTER SET 'utf8mb4' COLLATE 'utf8mb4_0900_ai_ci' NULL,
   PRIMARY KEY (`id`));
